articles table...refactoring

+++ resources/js/dashboard/article/ArticlesIndex.js
@@ -2,12 +2,13 @@ import React from 'react'
 import { Api, Routes, ApiCallHandler } from '../utility/URLs'
 import { Table } from 'react-bootstrap'
 import { Link } from 'react-router-dom'
+import ArticlesTable from '../components/ArticlesTable'
 
 export default function ArticlesIndex(props) {
     const [articles, setarticles] = React.useState([])
 
     async function setup() {
-        ApiCallHandler(async () => await Api.fetchArticles(),
+        ApiCallHandler(async () => await Api.fetchArticles({ with: ['category', 'pageWithoutContent'] }),
             setarticles,
             'ArticlesIndex setup',
             true
@@ -16,36 +17,17 @@ export default function ArticlesIndex(props) {
     React.useEffect(() => {
         setup()
     }, [])
-
     function deleteArticle(id) {
-        ApiCallHandler(async () => await Api.deleteArticle(id),
-            setup,
-            'ArticlesIndex deleteArticle',
-            true
-        )
+        if (confirm("هل تود فعلا حدف التصنيف " + id)) {
+            ApiCallHandler(async () => await Api.deleteArticle(id),
+                setup,
+                'ArticlesIndex deleteArticle',
+                true
+            )
+        }
     }
+
     return (
-        <Table striped bordered hover>
-            <thead>
-                <tr>
-                    <th>#</th>
-                    <th>العنوان</th>
-                    <th>العنوان</th>
-                    <th></th>
-                </tr>
-            </thead>
-            <tbody>
-                {
-                    articles?.map((article, index) => (
-                        <tr key={index}>
-                            <td> <Link to={Routes.articleShow(article.id)}> {article.id}  </Link> </td>
-                            <td>{article.page_id}</td>
-                            <td>{article.category_id}</td>
-                            <td onClick={() => deleteArticle(article.id)}>حدف</td>
-                        </tr>
-                    ))
-                }
-            </tbody>
-        </Table>
+        <ArticlesTable articles={articles} deleteArticle={deleteArticle} />
     )
 }
